fix: deleting image by name:tag, name (no tag) and digest

+++ pkg/client/image.go
@@ -10,6 +10,7 @@ import (
 	kclient "github.com/acorn-io/acorn/pkg/k8sclient"
 	"github.com/google/go-containerregistry/pkg/name"
 	"github.com/gorilla/websocket"
+	"golang.org/x/exp/slices"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 )
 
@@ -170,21 +171,37 @@ func (c *DefaultClient) ImageDelete(ctx context.Context, imageName string, opts 
 
 	image, err := c.ImageGet(ctx, imageName)
 	if apierrors.IsNotFound(err) {
+		if strings.HasSuffix(imageName, ":latest") {
+			// if the image is not found, we try to delete the image without the tag
+			return c.ImageDelete(ctx, strings.TrimSuffix(imageName, ":latest"), opts)
+		}
 		return nil, nil
 	} else if err != nil {
 		return nil, err
 	}
-	if len(image.Tags) == 1 {
-		return image, c.Client.Delete(ctx, image)
-	}
 	var remainingTags []string
 
-	imageParsedTag, err := name.NewTag(imageName, name.WithDefaultRegistry(""))
+	imageParsedRef, err := name.ParseReference(imageName, name.WithDefaultRegistry(""), name.WithDefaultTag(""))
 	if err != nil {
 		return image, nil
 	}
+	tagToDelete := imageParsedRef.Name()
+	if _, ok := imageParsedRef.(name.Digest); ok {
+		// if the image is referenced by digest, we need to delete the tag with only registry/repository
+		tagToDelete = imageParsedRef.Context().Name()
+	}
+
+	// Getting an image, auto-suffixed with :latest also returns images that don't have that tag (explicit :latest) at all, potentially yielding errors down the line
+	if !slices.Contains(image.Tags, tagToDelete) {
+		return image, fmt.Errorf("image %s not found", imageName)
+	}
+
+	if len(image.Tags) == 1 {
+		return image, c.Client.Delete(ctx, image)
+	}
+
 	for _, tag := range image.Tags {
-		if tag != imageParsedTag.Name() {
+		if tag != tagToDelete {
 			remainingTags = append(remainingTags, tag)
 		}
 	}
